bootleg reshuffle
