+++ src/mame/arcade.flt
@@ -256,6 +256,8 @@ copsnrob.cpp
 corona.cpp
 cosmic.cpp
 cps1.cpp
+cps1bl_5205.cpp
+cps1bl_pic.cpp
 cps2.cpp
 cps3.cpp
 cpzodiac.cpp
